server/requirements.txt
aniso8601==8.0.0
azure-cognitiveservices-vision-face==0.4.1
azure-common==1.1.25
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
emoji==0.6.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msrest==0.6.19
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
Pillow==8.0.1
pymongo==3.11.0
python-dotenv==0.15.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.11
Werkzeug==1.0.1
